Fix Table of contents/ in-page navigation block with slate
